aws-user-certs/requirements.txt
# ON WSL UBUNTU

# sudo apt-get install libssl-dev libffi-dev

# for running aws commands you need to get aws cli:

# curl "https://awscli.amazonaws.com/awscli-exe-linux-x86_64.zip" -o "awscliv2.zip"
# unzip awscliv2.zip
# sudo ./aws/install


pyOpenSSL